market-mind/requirements.txt
streamlit
transformers
torch
python-dotenv
